Minor warning fix

+++ dsptk/convolution.cc
@@ -25,7 +25,7 @@ namespace dsptk {
 
 	std::vector<double> convolve_out(const std::vector<double>& input, const std::vector<double>& kernel) {
 		if (input.size() == 0 || kernel.size() == 0) return std::vector<double>(0);
-		int resultSize = input.size() + kernel.size() - 1;
+		size_t resultSize = input.size() + kernel.size() - 1;
 		auto result = std::vector<double>(resultSize, 0.);
 
 		for (int i = 0; i < resultSize; i++) {
